feat: create new dynamic routing page
